+++ lucid/models/vision/mobilenet/__init__.py
@@ -6,18 +6,26 @@ from lucid.models.vision.mobilenet._model import (
     MobileNetV1ForImageClassification,
 )
 from lucid.models.vision.mobilenet._pretrained import (
-    mobilenet_v1, mobilenet_v1_cls,
-    mobilenet_v1_075, mobilenet_v1_075_cls,
-    mobilenet_v1_050, mobilenet_v1_050_cls,
-    mobilenet_v1_025, mobilenet_v1_025_cls,
+    mobilenet_v1,
+    mobilenet_v1_cls,
+    mobilenet_v1_075,
+    mobilenet_v1_075_cls,
+    mobilenet_v1_050,
+    mobilenet_v1_050_cls,
+    mobilenet_v1_025,
+    mobilenet_v1_025_cls,
 )
 
 __all__ = [
     "MobileNetV1Config",
     "MobileNetV1",
     "MobileNetV1ForImageClassification",
-    "mobilenet_v1", "mobilenet_v1_cls",
-    "mobilenet_v1_075", "mobilenet_v1_075_cls",
-    "mobilenet_v1_050", "mobilenet_v1_050_cls",
-    "mobilenet_v1_025", "mobilenet_v1_025_cls",
+    "mobilenet_v1",
+    "mobilenet_v1_cls",
+    "mobilenet_v1_075",
+    "mobilenet_v1_075_cls",
+    "mobilenet_v1_050",
+    "mobilenet_v1_050_cls",
+    "mobilenet_v1_025",
+    "mobilenet_v1_025_cls",
 ]